+++ R/as_aggregation_structure.R
@@ -3,10 +3,10 @@ as_aggregation_structure <- function(x, ...) {
 }
 
 as_aggregation_structure.default <- function(x, w = NULL, ...) {
-  aggregation_structure(list(x), w)
+  aggregation_structure(as.list(x), w)
 }
 
-as_aggregation_structure.agg_pindex <- function(x, w = NULL, ...) {
+as_aggregation_structure.aggregate_index <- function(x, w = NULL, ...) {
   eas <- x$pias$eas
   w <- if (is.null(w)) {
     rep.int(1, length(eas))
@@ -14,7 +14,12 @@ as_aggregation_structure.agg_pindex <- function(x, w = NULL, ...) {
     as.numeric(w)
   }
   names(w) <- eas
-  pias(x$pias$child, x$pias$parent, x$levels, eas, w, x$pias$height)
+  new_aggregation_structure(x$pias$child,
+                            x$pias$parent,
+                            x$levels,
+                            eas,
+                            w,
+                            x$pias$height)
 }
 
 as_aggregation_structure.data.frame <- function(x, ...) {
@@ -23,5 +28,5 @@ as_aggregation_structure.data.frame <- function(x, ...) {
 }
 
 as_aggregation_structure.matrix <- function(x, ...) {
-  aggregation_structure(as.data.frame(x))
+  as_aggregation_structure(as.data.frame(x), ...)
 }