+++ NEWS.md
@@ -8,6 +8,14 @@
 
 - The first argument to `elemental_index()` is now `x`, not `rel`, to be consistent with the rest of the functions.
 
+## New features
+
+- New functions `as_aggregation_structure()` and `is_aggregate_index()` to coerce (usually) tabular data for aggregation weights into an aggregation structure, and test if an object is an aggregation structure.
+
+- A method for `[[` for index objects to extract a single index value.
+
+- The weights for an aggregation structure can be replace with `weights(pias) <- vector`.
+
 ## Improvements
 
 - The `levels()` replacement function now gives an error for indexes and aggregation structures, rather than adding a levels attribute that does nothing.